Gemv: move functor-level interfaces to the top-level header
