+++ batched/dense/impl/KokkosBatched_Trsv_Team_Internal.hpp
@@ -8,7 +8,7 @@
 #include "KokkosBlas1_set_impl.hpp"
 #include "KokkosBlas1_team_scal_impl.hpp"
 #include "KokkosBatched_InnerTrsm_Serial_Impl.hpp"
-#include "KokkosBlas2_team_gemv_impl.hpp"
+#include "KokkosBlas2_gemv.hpp"
 
 namespace KokkosBatched {
